% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{format.pairrel}
\alias{format.pairrel}
\title{Format relationship description}
\usage{
\method{format}{pairrel}(x, cap = TRUE, includePaths = TRUE, ...)
}
\arguments{
\item{x}{An output of \code{\link[=verbalise]{verbalise()}}.}

\item{cap}{A logical indicating if the first letter of each path description
should be capitalised. By default TRUE.}

\item{includePaths}{A logical indicating if the complete paths should be
included in the output. By default TRUE.}

\item{...}{Not used.}
}
\description{
This is the main formatting function called by \code{print.pairrel()}.
}
