% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{habsburg}
\alias{habsburg}
\title{Habsburg pedigree.}
\format{
A \code{ped} object containing a pedigree with 29 members.
}
\source{
Adapted from \url{https://en.wikipedia.org/wiki/Habsburg_family_tree}
}
\usage{
habsburg
}
\description{
A subset of the royal Habsburg family, showing the ancestry of (the
infamously inbred) King Charles II of Spain.
}
\examples{

plot(habsburg, hatched = "Charles II", cex = 0.7)

verbalise(habsburg, ids = parents(habsburg, "Charles II"))

}
\keyword{datasets}
