% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbalise.R
\name{verbalise}
\alias{verbalise}
\title{Describe a pairwise relationship}
\usage{
verbalise(x, ids, verbose = TRUE)
}
\arguments{
\item{x}{A \code{ped} object.}

\item{ids}{A vector of length 2, containing the names of two pedigree members.}

\item{verbose}{A logical.}

\item{level}{A nonnegative integer (for internal use).}
}
\value{
A character vector.
}
\description{
Describe a pairwise relationship
}
\examples{

# Example 1: Family quartet

x = nuclearPed(2)
verbalise(x, 1:2)
verbalise(x, 2:3)
verbalise(x, 3:4)

# Example 2: Complicated cousin pedigree

y = doubleCousins(degree1 = 1, removal1 = 1, half1 = TRUE,
                  degree2 = 2, removal2 = 0, half2 = FALSE)
verbalise(y, leaves(y))
# Example 3: Full sib mating

z = fullSibMating(1)
verbalise(z, 5:6)
verbalise(z, c(1,5))

}
