% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbalisr-package.R
\docType{package}
\name{verbalisr-package}
\alias{verbalisr}
\alias{verbalisr-package}
\title{verbalisr: Describe Pedigree Relationships in Words}
\description{
Describe in words the genealogical relationship between two members of a given pedigree, using the algorithm in Vigeland (2022) \doi{10.1186/s12859-022-04759-y}. 'verbalisr' is part of the 'pedsuite' collection of packages for pedigree analysis. For a demonstration of 'verbalisr', see the online app 'QuickPed' at \url{https://magnusdv.shinyapps.io/quickped}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/magnusdv/verbalisr}
  \item \url{https://magnusdv.github.io/pedsuite/}
  \item Report bugs at \url{https://github.com/magnusdv/verbalisr/issues}
}

}
\author{
\strong{Maintainer}: Magnus Dehli Vigeland \email{m.d.vigeland@medisin.uio.no} (\href{https://orcid.org/0000-0002-9134-4962}{ORCID})

}
\keyword{internal}
