% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbalisr-package.R
\docType{package}
\name{verbalisr}
\alias{verbalisr}
\title{verbalisr: Describe pedigree relationships in words}
\description{
Describe the genealogical relationship between two members of a given
pedigree. verbalisr is part of the \emph{ped suite} collection of packages for
pedigree analysis.
}
