% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.pairrel}
\alias{print.pairrel}
\alias{format.pairrel}
\title{Format and print relationship descriptions}
\usage{
\method{print}{pairrel}(x, ...)

\method{format}{pairrel}(
  x,
  cap = TRUE,
  simplify = FALSE,
  abbreviate = FALSE,
  collapse = NULL,
  includePaths = !simplify,
  ...
)
}
\arguments{
\item{x}{An output of \code{\link[=verbalise]{verbalise()}}.}

\item{...}{Arguments passed on to \code{format.pairrel()}.}

\item{cap}{A logical indicating if the first letter of each path description
should be capitalised. By default TRUE.}

\item{simplify}{A logical. If TRUE, the descriptions of lineal and avuncular
relationships are simplified. Default: FALSE.}

\item{abbreviate}{A logical. It TRUE, various abbreviations are applied to
the descriptions, e.g. 'great-great-' -> 'gg-' and 'once removed' -> '1r'.
Default: FALSE.}

\item{collapse}{A single string, or NULL. If given, and the relationship has
multiple descriptions, these are concatenated with \code{paste(..., collapse = collapse)}.}

\item{includePaths}{A logical indicating if the complete paths should be
included in the output. By default TRUE.}
}
\description{
This documents the options for formatting and printing the output of
\code{\link[=verbalise]{verbalise()}}.
}
